Create option to switch body from the top
If fromTop is set to true, bodies will treat the top left corner as an
origin when switching, rather than the bottom left.

Body switching from the top is implemented where the legs disappear upon
taking flight. This is so that the legs go up into the body, rather than
the body moving down suddenly (causing a jerky flight).

[[WARNING: Untested!]]