Use LevelManager in LevelScreen and LevelTitleScreen to change Levels
LevelTitleScreen calls level_notitle because it switches levels without
showing the title screen first, whereas LevelScreen just wants to switch
to a different screen.

In the future, LevelScreen will not handle which level is switched to.
It does right now because a mechanism has not been created which allows
LevelManager to manage a sequence of levels.