Create FreeTypeFontLoader and FreeTypeFontParameters
Having a type that extends AssetLoaderParameters is necessary in order
to pass parameters to the font generator. This type is
FreeTypeFontParameters.

FreeTypeFontParameters simply contains a public FreeTypeFontParameter
because there is not a way to extend both FreeTypeFontParameter and
AssetLoaderParameters. Neither are interfaces.

Color and scale are not set within the loader, because these can be set
on a BitmapFont after it has been generated. The opposite is true for
anything that goes in a FreeTypeFontParameter.