Update FreeTypeFontParameters so that it works with BitmapFontGroup
Previously, FreeTypeFontParameters only held a single
FreeTypeFontParameter object. We are working towards having a
FreeTypeFontLoader that can load multiple BitmapFonts from a single
FreeTypeFont, so FreeTypeFontParameters needs to be able to hold
multiple sets of parameters.

The name that is associated with any FreeTypeFontParameter will be the
name that is used to get the BitmapFont that is created from that
FreeTypeFontParameter from its BitmapFontGroup.